[ADD]search in shop page, product page template

+++ ecommerce/products/models.py
@@ -13,10 +13,11 @@ class Category(BaseModel):
 class Product(BaseModel):
     product_name = models.CharField(max_length=100)
     category = models.ForeignKey(Category, on_delete=models.CASCADE, related_name="products")
-    price = models.IntegerField()
+    price = models.FloatField()
     product_description = models.TextField()
     slug = models.SlugField(unique=True, null=True, blank=True)
     product_image = models.ImageField(upload_to="product", null=True)
+
 
 
 class ProductImage(BaseModel):
